requirements.txt
numpy
pandas
pillow
scikit-image
tqdm
zipfile36


tensorflow-gpu
keras==2.1.6

